Loading test
Game generates a 10x10 sprite map. When moving 2 units, a new map is generated in a new place. Old tiles occupying new space are moved to a new place, while the empty places are filled with new cells.

This is just a test - sprites will be replaced by tilemaps later.