Tile sprites
Instead of  stamping blank squares on map, game now can stamp Textures2D.

This should work, but there is a bug which causes a nullReference while loading textures, even tho the variable in question is asigned in the inspector and does not change at runtime.

Will fix that after noon - I'm tired.