Biome update
Added more biomes and tile types. They generate alright, but the code responsible for picking specific tile types, and their graphics, should be tweaked a bit more - to be both random, and to take closeness to water into consideration.

More tile objects should also be added.

Additionaly, added a functionality, that generates the world not on start, but when space bar is pressed. This can be useful for pre-game options.

These options can also include render settings, which have been added (but not yet expanded) in the world manager class.